requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
cloudinary==1.44.0
dj-database-url==2.3.0
dj-rest-auth==7.0.1
Django==5.2
django-allauth==65.7.0
django-browser-reload==1.18.0
django-cloudinary-storage==0.3.0
django-contact-form==5.2.0
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
gunicorn==23.0.0
idna==3.10
packaging==24.2
pillow==11.1.0
psycopg2==2.9.10
PyJWT==2.9.0
python-dotenv==1.1.0
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.13.1
urllib3==2.4.0
whitenoise==6.9.0
